Replace literal return values with GradleValue objects.
With this, the users of the GradleBuildModel api can get additional
information about a value like the PsiElement corresponding to the
value and any variables resolved to obtain a perticular value.

Previoulsy, only ExtModel and ArtifactDependencyModel classes were
returning GradleValue objects, now most of the other model classes
are doing the same.

This is the main change in the process of adapting the GradleValue
return type. Some more changes will follow this to replace the list
and map values with the list and map of GradleValue objests.

Change-Id: I777f35ff91072d6c2d7a67998659f402a694a13d