pkginfo: fix race condition by serializing fetches
Previously, canceling the dialog and opening it again during feching
would cause issues e.g. while the second run was in progress, the first
one would finish and uninstall the APK, or kill the forward etc.
This iss not necessarily optimal (we could e.g. rewire the listener
to get the data sooner, or we can terminate the first run early etc)
but should work correctly and is a common case.

Change-Id: I4a628806b16b30d6e7d8ed9afe4d95e2f6a9c63e
Fixes: 31699123