Fixed detection of Gradle projects.
Before this CL, Projects#isBuildWithGradle(Project) considered
a project to "be built by Gradle" if at least one of the modules
in the project has the AndroidGradleFacet. This is wrong,
because a Gradle project may have zero modules (for example,
when the settings.gradle file is empty.)

Because of this bug, if a project has an empty settings.gradle
and even though Gradle sync was successful, Android Studio
considered such project a non-Gradle project. The side effect
is that the "Gradle Sync" toolbar action disappeared.

The fix is, in addition to the existing check, verify that there
was a Gradle sync, by checking the timestamp of the last sync. If
the timestamp is not -1, a Gradle sync was executed, which means
that the project is a Gradle project.

Besides this bug, this CL fixes a similar bug where the IDE
will consider a Gradle project with no modules to be a "legacy
IDEA project" and it will offer to auto-detect "Android framework".
This validation is in AndroidFrameworkDetector. To fix this
bug we need to apply the same check mentioned above. Checking for
a sync timestamp is not enough, because framework detection may
happen before sync. To overcome this issue, we also check that
the project has a top-level build.gradle file.

To make code more testable, this CL introduces GradleProjectInfo,
a project-level service that is easily mockable in tests. This
class will eventually replace Projects. Projects makes testing
difficult because:
1. It contains way too many methods
2. All methods are static

Fixes
https://code.google.com/p/android/issues/detail?id=203384

Change-Id: I5e2dfd4ea2b3358d73972f68d6c42e596225587b