Added tests to ensure that the IDE attempts to stop a build.
This CL includes a small refactoring and tests to
ensure that the IDE is sending Gradle the signal to
stop a build.

This is related to
https://code.google.com/p/android/issues/detail?id=201098

This CL does not fix the bug, but at least we can
confirm the bug is not in the IDE. We are working
with Gradle on fixing this bug.

Change-Id: Ibde2415d221d84434546efc03ea2f4be4cfebb3e