Splitting android plugin definition out of android module
Currently it is not possible for any other module within the android
plugin to depend on any code in the android module, since the android
module is in effect the root module of the project.
Splitting out the plugin definition itself (and
AndroidStudioInitializer) into a separate module which can then depend
on others (which can then depend on "android").
Essentially, this was renaming "main_AndroidStudio" to "android-plugin"
and moving some files between the two.

In order to move AndroidStudioInitializer out it was necessary to move
some static logic to other places.

Change-Id: I27688afd3ebb85bab5a00bd0223d9d523903a840