Fixed transitive dependencies in Java modules.
The bug was that inter-module dependencies in Java
modules were not set as 'exported', resulting in
transitive module dependencies not being visible,
for example:

  app -> lib1 -> lib2

'app' module was not able to see 'lib2' because
'lib1' did not set the dependency to 'lib2' as
'exported'.

This CL also replaces the old module customizers
with the new setup steps that work with both old
and new Gradle sync.

Fixes
https://code.google.com/p/android/issues/detail?id=225923

Change-Id: I6f24ce05f49c7a5618cf90af77b6d792f86b7e68