Fix webp preview delta
The code which shows a preview of the delta between the PNG and WEBP
images was using a hardcoded alpha of 255. The problem with that is
that the encoder feels free to ignore the RGB pixels when the alpha
value is 0 - so even for lossless images there would be some big
visual noise in the delta image. Tweak this to (a) apply the encoded
alpha value instead of a hardcoded alpha value, and for lossless
images we can skip the whole computaiton.

Change-Id: I456785f474a8f7997d58fdd6fc03bae0942f4a34