Fix exception when switching profiling process
This change severs the dependency on ProfilerEventListener
in the DataStore logic.

Instead of having the indivudal Pollers fire events to notify
that the profiler service has been disconnected, the monitor
now runs a heart beat poller to check if perfd is still alive.

Added additional handling around device/client start/stops
to pause the profilers instead of destroying everything.

Test: refactoring CL/manual. Tests to mock profiler server
connect/disconnect to come down the road.

Change-Id: Ib68a15a1afabc4e89dea46aba61e9e740c79d0cd