Add final modifier to method in BrowserHelpAction
BrowserHelpAction, which used to be a final class, was made non-final,
so that GPU Help Actions could inherit from it. However, there's no
reason for its actionPerformed method to not be final, since no classes
override it at the moment (and we can always remove the modifier
later if it ever becomes necessary to do so).

Test: No testing required, as we are simply adding a final modifier
Change-Id: I7d2f343a61429436eeea32b13af629fa4089a621