Fixing search scopes in JUnit and Android Tests
When double clicking a failed UnitTest or AndroidTest it often takes
user to the wrong file if there are both Android and Unit test classes
with the same name. It happens because the GlobalSearchScope is not
correct.

This CL fixes AndroidTest and JUnit scopes by implementing the method
initScopes() in both AndroidTestConsoleProperties and
AndroidJUnitConsoleProperties.

Test: com.android.tools.idea.run.testing.DuplicateClassNameTestScopesTest
This change is tested by creating an android application with two
different classes with the same name (in different test scopes) and
verifying if the locator finds only the correct one.

Fixes bug: http://b.android.com/226064

Change-Id: I6c46e6ad40133beada501a9bc268d22b15d8169b