NewProjectTest: Fix timeout waiting for ideFrame
When creating a new project (from the welcome screen) there is a
"Gradle building" modal dialog, before the project IDE is open.
Tests were timing out because they wait 10s for the frame to be shown
but gradle some times takes a bit longer to build.
Adding a call to waitForBackgroundTasks() will make sure we only
start looking for the Frame after the modal dialog is dismissed.
Failure was about 1 in 5.

Test: NewProjectTest
Change-Id: If79d69647441e50b15916a156dae5ae7f2e08887