Splitting out designer-tests and sdk-updates-tests.
The problem is that when running UI tests through IJ, we have to have
everything that's included in production on the classpath. This can
exclude test-only modules, and, in the case that that module has a
plugin.xml that replicates some of the plugin.xml in android-plugin, it
must be excluded.
When running tests through ant, test modules must be included
(obviously, since we need to run the tests). But we can prune those out
via the gant script.

This change splits the tests (including the plugin.xml) out of designer
and sdk-updates, so that those modules can be excluded when running ui
tests through IJ.

Test: This should allow existing UI tests to run through IJ. No
additional tests are needed/possible.

Change-Id: I330d5e5319a6025008c963ee9806f64b7535ffce