Fixing JUnit run configuration from package
When creating an AndroidTestRunConfiguration for the entire package
(with right click on it) and after creating a JUnit run configuration
the same way, the JUnit configuration was not created because the IDE
considered it as a duplicate of the existing Android configuration.

This CL fixes this problem by verifying if the configuration being
created concerns the same scope as the existing one.

Test: correct behavior assured by
AndroidTestConfigurationProducer#testConfigurationsAreDifferent

Fixes bug: http://b.android.com/215255

Change-Id: Ia19001592161cbea8267ad7614354638fd600f7e