227854: Fix layout rendering with 2.3.0-alpha2
This CL fixes layout rendering with support libraries when
using Gradle plugin 2.3 canary 2 or later. Now that the build
cache is enabled by default, resource files are ending up in
a .build-cache folder instead of the normal exploded-aar folder,
and various parts in the layout editor were relying on this
location to compute group-id/artifact-id in order to deduplicate
library dependencies.

Test: A regression test for this will be done in a follow-up
CL; this needs to go in ASAP to fix the blocker. Tested manually.

Change-Id: I8cdc6c5cc894e50eb2b9cab665ffc66ce8efb196