Fix line reducer assertion errors
The data pollers should be responsible for clearing their data between
when polling stops and then starts again, but before now they weren't
doing it. This meant that you could occasionally start monitoring a new
process while the poller still had old data leftover from a previous
process.

Test: Tested manually. Before, you would get this assertion 100% of
the time if you killed an app and relaunched it. After this fix,
the assertions don't happen anymore in that case.

Change-Id: If42bab6178c5c193e80ce2268452de9c5532eb7d