Add Gradle Sync Performance Test
Add a new test class GradleSyncPerfTest, which would try to load
project external/android-studio-gradle-test, and save event logs
to trk file.
This test resides in package com.android.tools.perf.idea*, so that
it is not executed on post submit buildbot. Instead, there will be a
new sync perf buildbot to run this test continously on post submit.
It is also excluded from bazel bot.

Test: covered by existing unittest.

Change-Id: I87dfe693308e39e8585c9744ef84d3a8f62890c3