Changed LineChart + AxisComponent to snap during the first update
- Fixed an issue where axes were registered to the Choreographer before
  the LineChart in the monitors.

Test: Added AxisComponentTest + LineChart tests
This change does not affect most existing ImageDiffTests as they all have
initial ranges that are greater than any y values in the data series. Re-enabled
two tests that were commented out during the initial creation of the adt-ui-model
module. One test has accumulated an error of 0.7% likley due to changes in the
interpolation ratio, so it has been recorded.

Change-Id: If3b78e8d97598a04a767863f21ce1624174d96a9