Moves Capture and CaptureView to its own classes.
This cl creates the concept of CpuCapture. When the CPU stage
has a capture a new view is created for it, and it's displayed on
screen.

This separates the responsibilities of L2 and L3 rendering in two
different objects, making it easier to add new renderings of the same
capture.

Change-Id: I5e05bf58a2b49dbed0e6fd180761b324fedb4870
Test: to follow up