Update IR experimental UI option #3 (Re-install and Run)
This CL changes the experimental UI option #3 from "Stop and Run" to
"Re-install and Run". The main changes are:
 - the icon has now changed to be more visible.
 - it is always enabled. If there is no active session, then clicking on
   it will be equivalent to clicking Run, but with installing an APK.
 - when this is enabled, the hot & warmswap notifications will have a
   little bit of text appended to indicate that all results may not be
   visible.

This CL inlines the UpdateActionBase class directly into the
HotswapAction class since the base class is not used anymore in the
Re-install action.

Test: none, this is experimental code for a UI test that is hidden under
a flag, will likely be deleted after some UI testing

Change-Id: Id93dddddf3e247a811d0591af757873bbdc6fbf2