Adds TopDown tree view for CPU captures
This CL adds a top-down view of the CPU capture. The HNode
tree is processed merging all the calls to the same method from the
same parent.

When the selection range changed, the delta ranges between the
previous range and the current is calculated. These little ranges
are then used to identify which nodes in the tree need to be updated.

The top-down tree is then traversed modifying those nodes that
overlap with the delta ranges, and only those.

Test: Unit tests included. Yay!
Change-Id: I5806852fc035d3a38979b94cf64e1ceda9f855b3