Now sync messages from previous sync are removed on sync error.
This CL is the second part fix for
https://code.google.com/p/android/issues/detail?id=226780

The issue was that the sync messages from the previous sync
were not removed if the current sync failed due to an exception
thrown either by Gradle or the Android plugin.

The fix is to clear any existing sync messages when sync
fails, before showing any messages related to the new failure.

Test: JUnit tests.
Change-Id: Ida564f3f3715dc362e5da452eeb7261192dddd42