New Project Test: Facet Disposed Exception on NlModel
Call to updateModel() is slow, and the facet may be disposed while
the method is running. If there is an error, don't log it if the facet is
already disposed.

com.intellij.util.IncorrectOperationException: Sorry but parent:
Android (app) has already been disposed (see the cause for stacktrace)
so the child: com.android.tools.idea.res.AssetRepositoryImpl@53fdc0bb will
never be disposed
	at com.intellij.openapi.util.objectTree.ObjectTree.register(ObjectTree.java:65)
	at com.intellij.openapi.util.Disposer.register(Disposer.java:92)
	at com.intellij.openapi.util.Disposer.register(Disposer.java:86)
	at com.android.tools.idea.res.AssetRepositoryImpl.<init>(AssetRepositoryImpl.java:44)
	at com.android.tools.idea.rendering.RenderTask.<init>(RenderTask.java:166)
	at com.android.tools.idea.rendering.RenderService.createTask(RenderService.java:238)
	at com.android.tools.idea.uibuilder.model.NlModel.inflate(NlModel.java:362)
	at com.android.tools.idea.uibuilder.model.NlModel.updateModel(NlModel.java:425)
	at com.android.tools.idea.uibuilder.model.NlModel$2.run(NlModel.java:272)

Test: NewProjectTest@scrollingActivityFollowedByBasicActivity()
Change-Id: I404e3ae127b78b5fee3bd7d285b113a509d7d2c1