Handle ProjectResourceRepo root change invalidation async
Otherwise there can be a deadlock. This way the
AWT thread drops the write lock and parsing work
can continue with read locks.

Also, if there are many new roots for each module
then this avoids heavy parsing on the AWT thread.
Ideally, we would peek at the new roots, and kick
off a multi-threaded parsing task, but that would
involve a private API into the folder manager.

Bug: b.android.com/227845
Bug: b/32720280
Test: added a "normal" root change unit test
Didn't actually attempt to hit the deadlock

Change-Id: I37a8c58095be1ef2d18e633b15b66b29c07a3fc8