Moves logic for closest point of data series to one place.
We now use getDataForRange(x, x) as a generic way to get the value
of a point. Given that series add the padding needed to get render
this range, we can get the value we need without specializing per
class.

This actually fixes CPU Label which was incorrectly showing 0% most
of the time.

Test: Refactoring
Change-Id: I81ac53ebdfbfeebcc9213d3b5e1e7b6166aabff2