Refactor TabularLayout logic
Before, essentially the same logic, slightly tweaked, appeared in
three places in the previous iteration of this layout manager -
getting the preferred size, getting the minimum size, and doing
the layout all shared a majority of code.

This approach refactors much of the common logic out into
the SizeCalculator and LayoutResult inner classes.

Besides removing repeated code, this change also paves the way
for letting us set sizing rules on rows, not just columns,
in the next CL.

Test: Internal refactoring. TabularLayoutTests still pass and new
code is fully covered.

Change-Id: Id01966dd3a7903fafb03d037fa52b54a4970fbc0