Fixes going back to CPU profiler after a capture
CpuCaptures are now effectively a cache of a parsed trace.
When a new trace needs to be displayed we get it with
getTrace. The flow of coming back to a trace still
needs to be worked out and also the performance of
parsing a file, but this CL fixes the NPE we were
getting and takes us closer to the final solution.

Change-Id: Ifb88ca500d0ad1786accbabb5caf4b7b0e0c92f8
Test: To follow up.