Only remove libraries not needed after sync.
Previously, Gradle sync will remove all libraries
from the project before performing a Gradle sync.
The purpose for this action was to eliminate the
chance of ending up with libraries that were not
needed. For example, if the user removed 'Guava'
as dependency and synced the project, the project
after sync should no longer have 'Guava' as
dependency.

The problems with this approach are:
1. From the moment sync started until it ended,
   editors will not be able to resolve references
   to classes in libraries that were removed
   before sync, resulting in poor user experience.
2. Removing libraries triggered indexing, which
   competes with Gradle sync for I/O.

This CL removes the need to remove libraries
before sync. Instead, it keeps track of the
libraries that are going to be used after sync
and only removes the ones that are not used.

Fixes:
https://code.google.com/p/android/issues/detail?id=193126

Test: JUnit tests.
Change-Id: Ia3bd1a72b42b84860381e1b21b7396cfb58f2db7