Don't refresh to find file when retriving API source in debugger.
Issue: getPsiFileByLocation tries to find the source file for android
sdk files. If the file is not already seen by the VFS, it calls
fileSystem.refreshAndFindFileByIoFile while on a non-EDT that causes the
exception.

This CL disables the synchronous refresh to avoid the exception. There
is no regression in behavior. It'll return the file instead throwing an
exception if the source file is not available.

Fixes: https://code.google.com/p/android/issues/detail?id=224667
Test: Added unit tests for getSourceForPsiFile
Change-Id: I907e4847a72a774e198a33ec73ed1ae19a19fced