Cleaned up facet removal during sync.
This CL ensures that facets get removed when a missing
Gradle model is not present. For example, if a module
has an AndroidFacet before syncing, but after sync that
module does not have an AndroidProject, it means that
it is no longer an Android module, and the facet should
be removed. The same applies to NdkFacet, JavaFacet and
GradleFacet.

Test: JUnit test
Change-Id: I5c1f8b2aecf580224cba0a51d404e2fdf175ed2f