Restore selection and open groups in property table.
When a property is changed (in the property table or the design
surface) the table will get a notification about a new component
with a new set of properties. If we are just changing one thing
we do NOT want an open group to collapse or the selection to
dissapear.

Test: Updated to cover new functionality.
Change-Id: If804460fd309cd154552208338a7e88010393e4f
(cherry picked from commit 45b913db93ae07750941125692c6b08183f8f4a7)