Fix custom JDK handling in project structure
If useEmbeddedJdk is disabled, currently PSD takes into account only
the JDK paths selected explicitly by the user via file chooser dialog.
If the path is input directly into the text field, it is not used during
validation, nor makes it the way to IDE settings after applying, which
is rather confusing and is not consistent with how Android SDK and NDK
selectors behave.

This CL addresses that, while still preserving the user-selected JDK
logic which improves user experience when checking/unchecking
useEmbeddedJDK checkbox (the path doesn’t get reset to empty value
all the time), and when launching the file chooser dialog (the suggested
directory will point to the current user-selected JDK, rather than
embedded JDK path, home path or current text field value).

Test: PSD can only be tested with a UI test, and there is currently no
fixture for doing so. This is to be followed up
Bug: b.android.com/228571

Change-Id: Iad0d9085f709aa2af3aa753601bb26f1e2ed76c6