CellRenderer for PTable should not keep state.
Currently a cell renderer caches the table and row for use in
displaying the correct buttons for the BrowsePanel.
The BrowsePanel shows the resource button and the design button.

Keeping this state could lead to wrong rendering. And since the
renderes are kept in a global variable we keep a reference to the
last property table displayed. Both are really bad.

In this CL the cached state state is removed. The data needed for
display is calculated during getTableCellRenderComponent.

Similarly a table cell editor was keeping state that should be
removed when done editing, since the editors are cached for the
lifetime of the project.

Test: Added tests for the table cell editor.

Change-Id: I95d0d195a074b14151ea0977ab4a5bb4f71762cf