Refresh VFS earlier for ProjectResourceRepositoryTest temp dirs
Currently trying to refresh late as a "recovery"
step, but forgot to invalidate folder manager
first so that may not have an effect.

Try refreshing earlier, and see if we can remove
the "recovery" step, which is a bit of a hack.
Early refresh is more like what other test utils
do (like TemporaryDirectory.kt).

Change-Id: Iede76cbd320a33e96e07b6be87241bc5cca5f3fd
Test: modifying existing test setup
Bug: b.android.com/227339