Increase TopDownNode and TopDownTreeModel coverage
Increased the coverage of both classes to 100% by adding the tests
described below.

TopDownNode: added a test method to test the methods related to the node
duration, such as getTotal, getSelf and reset. Also added a method to
test the getters of the node's data, such as getMethodName and
getPackage.

TopDownTreeModel: removed a path that is not reachable anymore (a null
check in a field that now is annotated with @NotNull).

Test: TopDownTreeModelTest and TopDownNodeTest
Change-Id: I6e60ee964bda88db1335b4e5fe0ca0ad79d0f866