Make the minimum supported IR API level to be 21
When targeting devices lower than 21, we'll show a message saying that
Instant Run requires API level 21 or higher.

This CL also fixes a bug where we were requiring the installation of a
platform even for API levels on which IR wasn't supported.

The codebase still has places (InstantRunBuilder etc) that do a
redundant check for the API level. Those will be cleaned up in a
subsequent CL.

Bug: https://code.google.com/p/android/issues/detail?id=229355
Test: Existing Instant Run Manual tests, automated tests will be added
in a later CL

Change-Id: I29a7f2dbb3cdc9b640a60ce4e3b2babca15af546