Async loading of hprof files
This CL adds asynchronous loading and parsing of hprof files.
Hprof files are now loaded through a separate mechanism, and
the views will always be listening for the latest capture
being loaded (not only hprof files). The separate mechanism
is therefore a generic asynchonous capture loader that can
handle the loading of any CaptureObject.

Due to changes needed to support asynchronous loading, the
views' dependency on the model has also been expanded and
thusly refactored. This CL therefore creates a clean
separation between view and model.

Test: tests for selection and consumption of async loading

Change-Id: I20bf3ed7a90f93897758657965612e2cd2fdc995