229643: Part 2: Unable to add any constraints between UI controls
There was another dependency on the exploded-aar directory
layout which meant that it didn't read in the R.txt id table
when processing AAR libraries in the build cache. This resulted
downstream in the R class generator not producing R fields which
are required to load these classes.

Test: This is covered by UI tests (once we update them to
use Gradle plugin 2.3 instead of 2.2.1.)

Change-Id: Iec87000d0f1c472a0089d8d1fd53a5fc3a50b0e7