Now MissingCMakeErrorHandler loads SDK repos only when needed.
This CL fixes a bug where MissingCMakeErrorHandler would try
to load SDK repos on the UI thread when it is created, slowing
down Gradle sync on slow Internet connections.

The fix is to load SDK packages asynchronously when the user
clicks the "quick fix" hyperlink. If loading is successful,
the IDE will automatically install CMake and start a new
"Gradle Sync" operation.

Fixes:
https://code.google.com/p/android/issues/detail?id=229563

Test: JUnit test and manual test.

Change-Id: I15afc2cb55cca380d03cbda3b503276fabca9004