Fix HotswapAction text when we cannot determine device
When a device is not found from the current session, we used to assume
that the device was of API level 1. This eventually ended up being shown
in the hotswap action's description (saying that IR isn't supported when
deploying to API level 1). This CL fixes that issue: we now explicitly
distinguish between not knowing the device and the device API level
being too low. This fixes b.android.com/229700.

This CL also makes the hotswap action's text include its shortcut.

Test: IR manual tests by QA

Change-Id: I09faefffc51a48a768ec1ee4cbdb5abfaa72f5e7