Make AdView and MapView visible in the layout editor
Set a background and a minimum size so the users can see them when they
add them.
These controls usually require setup done in the onCreate which we don't
run in layoutlib. This tries to workaround that problem.

This is a temporary workaround just to make them visible. In the case of
MapView we want to add a static image to display as background.

Bug: http://b.android.com/219193
Test: Updated LayoutPsiPullParserTest and TagSnapshotTest
Change-Id: I08773796d6e5212dc0cdd425952d862a22e8d81b