Delay theme preview window checks until out of dumb mode
Detecting the best configuration relies on the
AppResourceRepository, which may take a while to
bring up, and happens on the AWT thread. If it
happens to early, then the resource indexer
(which speeds this up) will not have had a chance
to run yet.

Bug: b/33429298
Test: ThemePreviewTest#testPreviewAvailabilityDumbMode
checks that we recover after getting out of dumb mode

Change-Id: I8b93901b85c68856341a324891ab3782ce18eec1