229626: Don't use webp for launcher icons
The developer console requires PNG files for launcher icons.
We recently switched to using WEBP for launcher icons for default
projects if the minSdkVersion >= 18, but that means that as soon
as you try to publish the app it will be rejected.

This CL updates the convert to webp action (which recursively
converts PNG images in a folder) to skip launcher icons.
It also updates the resource chooser to no longer rely on
layoutlib to render webp images since we have a direct decoder
in the IDE.

Finally, it registers the lint warning which is emitted for
non-PNG launcher icons as an IDE inspection.

Test: New unit tests added
Change-Id: Ifc7264aaaa24296e834e67d205fadddd40e720c9