Forcibly kill app when doing a cold swap.
The Run/Debug buttons now force a cold swap (w/ split apks). Normally,
installing the apk forces the app to be killed. However, if the user had
not made any changes, we wouldn't have killed the app, which would make
the run/debug buttons be inconsistent (the app gets killed only if there
are changes).

This CL forcibly kills the app as long as we are doing a cold swap in
order to make this consistent.

Fixes http://b.android.com/229947
Test: IR manual tests by QA

Change-Id: I150088a411e10d2452af82639fb29eb8556e8229