DeployTargetPickerDialogFixture: make selectDevice more robust
Even though the method description claims to select the device
whose name *begins* with `text`, it was only searching for an
exact match. This meant that we would fail to select an emulator
which is already running as in this case it's name has additional
text appended (API level and such). I am going to need this
functionality in a test I am writing, which tries to deploy to an
already-running emulator.

Test: Existing tests still pass.
Change-Id: I04b963b639980c61dacc722a50d8c048029e6e89