Removed logging of "libraries not found" during sync.
This CL removes logging of "libraries not found" in
SyncLibraryRegistry. This logging was done under the
assumption that libraries exist only in one module.

If 2 or more libraries explicitly depend on "guava"
(for example), when setting up the first module,
the "guava" will be marked as used, but when setting
the second module, "guava" won't be found, and a
error message will be logged. Because "guava" was
already found and removed for the first module, not
finding "guava" for the second module is actually
correct.

Test: existing JUnit tests should pass.
Change-Id: I3abcd4c0e42f589a050f9c52963702a7213948d2
(cherry picked from commit 06625a6837d6ee4f5e8363ce0d8e89bad8460e11)