Moves ranged series to a custom model.
This allows the UI to refer to specific series directly as
they now surface all the way up.

This fixes the problem to configure LineCharts via Strings,
and enables charts with the same name as we had before.

This change also removes duplicated code where we were
creating RangedSeries in several places.

Test: All existing tests pass, this only moved code around.
Change-Id: I2b696bd009c36623d878a3e14f630cf955c318d9
Bug: 33673863