Aspect dependencies are all weak
Before this change the view objects where being kept
alive by the model. And whenever we created a new one
all the dependencies were set and never removed.

Another alternative for this change is to follow the
Disposer pattern and explicitly clean up
dependencies. You can see the approach here:
ag/1737767. It's much more involved and a lot
of room for errors.

The tests added navigate the object graph
making sure that when views are attached there
is no hard-link to any of the view classes,
therefore making sure that they can be collected
on their own.

Test: Unit tests
Bug: b/33464893, b/33676083
Change-Id: I3e9452ccd19845ca7a2c60c9a1f8bbc539968898