Look for source attachments only for new libraries.
With this CL, the IDE, during Gradle sync, will
only try to find library source attachments for
*new* libraries. In the past, source attachments
were looked up for every library, to be later on
be ignored for existing libraries (which should
have already source attachments.)

In order to accomplish this, this CL also
splits DependenciesSetup into
AndroidModuleDependencySetup and
JavaModuleDependencySetup. The reason is that
for Java modules the source attachments are already
known, and there is no need to attempt to
attach annotations.jar to a library.

In addition, this CL changes file paths to be
java.io.File instead of Strings.

Test: JUnit tests.
Change-Id: Ic5312b47a4fa81d6abedc3485942c45d0a4b0c9a