Removed method addListener(GradleSyncListener) from AndroidFacet.
AndroidFacet is already doing too much and the removed method
is just a simple shortcut that does not add too much value, compared
to invoking the method in GradleSyncState directly.

This CL also fixes a bug in DynamicResourceValueRepository where
removing a GradleSyncListener by registering as a child disposable
was done in the wrong order.

Test: Existing JUnit tests should pass.
Change-Id: I34fa16b4d197fcc7d594bbc1846373b2663c8210