Get rid of ResourceTypeInspection; only use lint inspections
This CL removes the ResourceTypeInspection (which wasn't just
enforcing the resource type annotations; it was enforcing *all* the
support annotations.)

There were two separate implementations of these checks: the resource
type inspection, as well as a duplicated implementation over in
lint. These were separate because the ResourceTypeInspection leveraged
a lot of IDE infrastructure, and that was not available to the lint
implementation. And they weren't combined because the IDE one was more
capable (did deeper analysis) whereas the lint one was able to be run
from Gradle.

However, I've now updated the lint implementation to have the same
capabilities as the IDE one (and I've transfered all the unit tests
from the ResourceTypeInspectionTest up to lint where they're all
passing).  Therefore the IDE inspection can be removed, which removes
a lot of code duplication (and a lot of work where each new support
annotation's analysis must be implemented twice.)

This CL also tracks some updates to the lint infrastructure which
allows lint errors to provide context info to quickfixes directly,
instead of having to have the fixes parse error messages to pull this
info back out (this was necessary for the permission-related
quickfixes, where we need more information for the quickfix than is
practical to pack into the error messages.)

Test: Mostly code deleted; the lint checks are already tested over
in the tools/base project. AndroidLintTest updated.

Change-Id: I6c8612fc055a843d5b05b748479c27037c5b3480