Fixed UI test ThemeSelectorTest.
Test was broken by
https://googleplex-android-review.git.corp.google.com/1752652

The test expects that the Module in ConfigurationManager is never
null, even if the ConfigurationManager is disposed. My CL was
setting the Module to null when ConfigurationManager is disposed
(a good practice) but I couldn't figure out a way to handle a
null Module. This is acceptable, as long as we don't leak the
Module after disposal of ConfigurationManager.

Test: UI test.
Change-Id: I58e72b2a8fddd99fd22f169d5c94dbed934d040e