Removed forced project refresh after source generation.
A project refresh was forced after source generation to
have the IDE resolved symbols from classes in AARs that
were in the Gradle build cache. Because now exploded
AARs are in the build cache, instead of a module's
'build' folder, the IDE's file watchers were not
notified when AARs were exploded, and as a result,
symbols in editors were not resolved.

With
https://googleplex-android-review.git.corp.google.com/1740896
AARs are exploded during sync, instead of the source
generation step, forcing a project refresh is no longer
necessary.

Test: JUnit test.
Change-Id: I7c1df1aaca68c987d7781a622b2e27509d027c98