Change the way we detect when we need to scroll property panel.
Problem:
We want to scroll the current control into view when we navigate
the property panel (inspector & property table) with the keyboard
via TAB and shift-TAB.
But we dont want to scroll when the user clicks on something.

The current solution is an ugly hack.
I propose a "better" hack where we look at the current AWT event
that caused the focus change.

Test: TBD in followup CL.
Change-Id: Ic0e34af8984f07e891f1b35988553af8697326f7