Support UAST in lint
Lint has switched over to using UAST for AST analysis.
This CL updates the IDE portion to track the API changes.
The VersionChecks class used to reside in lint, but now that
lint has ported it form PSI to UAST and we still need the PSI
version in the Android plugin, move it here.

Test: The new UAST portions are tested by lint in tools/base,
and the IDE integration is already tested by AndroidLintTest.

Change-Id: I70c9ccf64cae775dc2cc4e23c0066e33eb791093